+++ chromag/epochs/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""Handle properties of the spectral lines."""
+"""Handle changing properties of the instrument."""
 
 import datetime
 import os
@@ -10,15 +10,21 @@ import epochs
 
 DateValue = TypeVar("DateValue", str, datetime.datetime)
 
-EPOCHS_ROOT = os.path.dirname(os.path.abspath(__file__))
-EPOCHS_CFG = os.path.join(EPOCHS_ROOT, "epochs.cfg")
-EPOCHS_SPEC = os.path.join(EPOCHS_ROOT, "epochs.spec.cfg")
+
+ep = None
 
 
-ep = epochs.EpochConfigParser(EPOCHS_SPEC)
-ep.read(EPOCHS_CFG)
-
-
-def get(property_name, date: DateValue):
+def get(property_name: str, date: DateValue):
     """Get property value for a given datetime."""
+    global ep
+
+    # read the epochs files if it hasn't been initialized already
+    if ep is None:
+        epochs_root = os.path.dirname(os.path.abspath(__file__))
+        epochs_spec_filename = os.path.join(epochs_root, "epochs.spec.cfg")
+        ep = epochs.EpochConfigParser(epochs_spec_filename)
+
+        epochs_cfg_filename = os.path.join(epochs_root, "epochs.cfg")
+        ep.read(epochs_cfg_filename)
+
     return ep.get(property_name, date)